using block.json
