+++ .gitignore
@@ -1,12 +1,12 @@
 # Directories/files that may be generated by this project
 node_modules
 index.js.map
+/vendor
 
 # Directories/files that may appear in your environment
 .DS_Store
 *.log
 /yarn.lock
 .vscode
-
-/vendor/
 build
+.wp-env.json
