+++ 11-tabs-esnext/index.php
@@ -19,64 +19,7 @@ add_filter( 'render_block', 'gutenberg_examples_11_esnext_enqueue_conditionally'
  * the corresponding context.
  */
 function gutenberg_examples_11_esnext_register_block() {
-
-	if ( ! function_exists( 'register_block_type' ) ) {
-		// Gutenberg is not active.
-		return;
-	}
-
-	// automatically load dependencies and version
-	$asset_file = include plugin_dir_path( __FILE__ ) . 'build/index.asset.php';
-
-	wp_register_script(
-		'gutenberg-examples-11-esnext-editor',
-		plugins_url( 'build/index.js', __FILE__ ),
-		$asset_file['dependencies'],
-		$asset_file['version'],
-		true
-	);
-
-	wp_register_style(
-		'gutenberg-examples-11-esnext-editor',
-		plugins_url( 'editor.css', __FILE__ ),
-		array( 'wp-edit-blocks' ),
-		filemtime( plugin_dir_path( __FILE__ ) . 'editor.css' )
-	);
-
-	wp_register_script(
-		'gutenberg-examples-11-esnext',
-		plugins_url( 'frontend.js', __FILE__ ),
-		array( 'jquery' ),
-		filemtime( plugin_dir_path( __FILE__ ) . 'style.css' ),
-		true
-	);
-
-	wp_register_style(
-		'gutenberg-examples-11-esnext',
-		plugins_url( 'style.css', __FILE__ ),
-		array(),
-		filemtime( plugin_dir_path( __FILE__ ) . 'style.css' )
-	);
-
-	register_block_type(
-		'gutenberg-examples/example-11-tabs-esnext',
-		array(
-			'editor_script' => 'gutenberg-examples-11-esnext-editor',
-			'editor_style'  => 'gutenberg-examples-11-esnext-editor',
-			'script'        => 'gutenberg-examples-11-esnext',
-			'style'         => 'gutenberg-examples-11-esnext',
-		)
-	);
-
-	register_block_type(
-		'gutenberg-examples/example-11-tab-esnext',
-		array(
-			'editor_script' => 'gutenberg-examples-11-esnext-editor',
-			'editor_style'  => 'gutenberg-examples-11-esnext-editor',
-			'script'        => 'gutenberg-examples-11-esnext',
-			'style'         => 'gutenberg-examples-11-esnext',
-		)
-	);
+	register_block_type( __DIR__ . '/build' );
 }
 
 /**
@@ -86,9 +29,9 @@ function gutenberg_examples_11_esnext_register_block() {
 function gutenberg_examples_11_esnext_enqueue_scripts() {
 	wp_register_script(
 		'gutenberg-examples-11-esnext-frontend',
-		plugins_url( 'frontend.js', __FILE__ ),
+		plugins_url( 'tabs-frontend.js', __FILE__ ),
 		array( 'jquery' ),
-		filemtime( plugin_dir_path( __FILE__ ) . '/frontend.js' ),
+		filemtime( plugin_dir_path( __FILE__ ) . '/tabs-frontend.js' ),
 		true
 	);
 }
@@ -98,9 +41,9 @@ function gutenberg_examples_11_esnext_enqueue_scripts() {
  * to function.
  */
 function gutenberg_examples_11_esnext_enqueue_conditionally( $block_content, $block ) {
-	// if ( 'gutenberg-examples/example-11-tabs-esnext' === $block['blockName'] ) {
+	if ( 'gutenberg-examples/example-11-tabs-esnext' === $block['blockName'] ) {
 		wp_enqueue_script( 'gutenberg-examples-11-esnext-frontend' );
-	// }
+	}
 	return $block_content;
 }
