+++ 11-tabs-esnext/tabs-frontend.js
